extract chart
