+++ src/App.js
@@ -1,21 +1,12 @@
 import React, { useMemo, useState } from "react";
 import { useQuery } from "react-query";
-import {
-  LineChart,
-  Line,
-  CartesianGrid,
-  XAxis,
-  YAxis,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-} from "recharts";
 import Select, { createFilter, components } from "react-select";
 import { FixedSizeList as List } from "react-window";
 
 import { queryOptions } from "./constants";
 import { getData, processData } from "./utils";
 import Table from "./Table";
+import Chart from "./Chart";
 
 const MenuList = ({ options, children, maxHeight, getValue }) => {
   const height = 35;
@@ -158,39 +149,5 @@ function App() {
     </div>
   );
 }
-
-const Chart = ({ data, counties, UIDs }) => {
-  const d = [...data].reverse();
-  const COLORS = ["#0088FE", "#00C49F", "#FFBB28", "#FF8042"];
-
-  return (
-    <div className="m-auto max-w-screen-xl">
-      <h1 className="text-2xl">Active Cases %</h1>
-      <ResponsiveContainer minHeight={400} width="100%">
-        <LineChart data={d}>
-          <CartesianGrid strokeDasharray={"3 3"} />
-          <XAxis dataKey={`${UIDs[0]}.date`} />
-          <YAxis
-            dataKey={`${UIDs[0]}.activePercent`}
-            // scale={"log"}
-            // domain={[0, "dataMax"]}
-          />
-          <Tooltip contentStyle={{ backgroundColor: "#333" }} />
-          <Legend />
-          {UIDs.map((uid, i) => (
-            <Line
-              key={uid}
-              dot={false}
-              name={counties[uid].Admin2}
-              dataKey={`${uid}.activePercent`}
-              unit="%"
-              stroke={COLORS[i % UIDs.length]}
-            />
-          ))}
-        </LineChart>
-      </ResponsiveContainer>
-    </div>
-  );
-};
 
 export default App;
